### 2022-06-16 #4481 @editorialbot
[PRE REVIEW]: tidyqpcr: Quantitative PCR analysis in the tidyverse.
**Submitting author:** @ewallace (<a href="http://orcid.org/0000-0001-8025-6361">Edward Wallace</a>)
**Repository:** https://github.com/ropensci/tidyqpcr
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
