### 2022-06-16 #4481 @ewallace
Hello and thank you.

The software was already reviewed and accepted at @ropensci, see review thread https://github.com/ropensci/software-review/issues/470

Suggestions from your list of reviewers & editors:
- editors lpantano, fboehm, Bisaloo, majensen, jni, csoneson all look appropriate
- this is an R package to analyse qPCR (molecular biology / clinical detection) data, aimed at researchers, so someone with R and gene expression / molecular biology / PCR expertise would be ideal
- seaaan already did a pre-review of the software in 2020, which has evolved considerably since then
- standage, iimog, tijeco, have the right keywords
