### 2022-06-17 #4481 @ewallace
Thank you @arfon.

Regarding the substantial scholarly effort criterion, we have cited the software in our preprint:
> Limitations of composability of cis-regulatory elements in messenger RNA
> Haynes S, Auxillos J, Jain A, Alibert C, Danecka W, Wallace EWJ, bioRxiv (2022).
> http://dx.doi.org/10.1101/2021.08.12.455418

And we expect to cite the software again.
