### 2022-06-22 b6b0f2a
rOpensci acceptance edits, addresses #188 (#189)
* DESCRIPTION for rOpensci

* News out of README.md

* Create NEWS.md

* Links to ropensci in README.md

* ropensci in codemeta.json

* Create codemeta.json

* Addressed CRAN gotchas

See https://devguide.ropensci.org/building.html#crangotchas

* NEWS.md link from README

* reword code of conduct lines in README

As described in https://devguide.ropensci.org/collaboration.html

* Update CONTRIBUTING.md

* Removed duplicated development

Co-authored-by: Samuel Joseph Haynes <37002508+DimmestP@users.noreply.github.com>
