+++ README.md
@@ -1,10 +1,10 @@
 <!-- badges: start -->
 [![Project Status: Active – The project has reached a stable, usable state and is being actively developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
-[![Codecov test coverage](https://codecov.io/gh/DimmestP/tidyqpcr/branch/main/graph/badge.svg)](https://codecov.io/gh/DimmestP/tidyqpcr/branch/main)
+[![Codecov test coverage](https://codecov.io/gh/ropensci/tidyqpcr/branch/main/graph/badge.svg)](https://codecov.io/gh/ropensci/tidyqpcr/branch/main)
 [![fair-software.eu](https://img.shields.io/badge/fair--software.eu-%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8F%20%20%E2%97%8F-green)](https://fair-software.eu)
 [![CII Best Practices](https://bestpractices.coreinfrastructure.org/projects/5287/badge)](https://bestpractices.coreinfrastructure.org/projects/5287)
 [![CRAN_Status_Badge](https://www.r-pkg.org/badges/version/tidyqpcr)](https://cran.r-project.org/package=tidyqpcr)
-[![R-CMD-check](https://github.com/ewallace/tidyqpcr/workflows/R-CMD-check/badge.svg)](https://github.com/ewallace/tidyqpcr/actions)
+[![R-CMD-check](https://github.com/ropensci/tidyqpcr/workflows/R-CMD-check/badge.svg)](https://github.com/ropensci/tidyqpcr/actions)
 <!-- badges: end -->
 
 # tidyqpcr - Quantitative PCR analysis in the tidyverse.
@@ -22,7 +22,7 @@
 	* [Installing tidyqpcr](#Installing-tidyqpcr)
 	* [Using tidyqpcr](#Using-tidyqpcr)
 * [Status](#Status)
-    * [News](#News)
+    * [News - see NEWS.md](NEWS.md)
 * [Features](#Features)
 	* [Current features](#Current-features)
 	* [Future priorities](#Future-priorities)
@@ -73,17 +73,17 @@ Next, you need a working installation of [Rtools](https://cran.r-project.org/bin
 
 Jeffrey Leek made [slides on installation and testing of Rtools](http://jtleek.com/modules/01_DataScientistToolbox/02_10_rtools/).
 
-### For all R users
+### Install via devtools (for all R users)
 
 Install the devtools R package, see [devtools installation instructions](https://www.r-project.org/nosvn/pandoc/devtools.html). 
 
 ```
 library(devtools)
-devtools::install_github("ewallace/tidyqpcr",build_vignettes = TRUE) ## Vignettes require cowplot package
+devtools::install_github("ropensci/tidyqpcr", build_vignettes = TRUE)
 
-## Alternatively, install without building the vignetttes to remove cowplot dependency 
+## Alternatively, install without building the vignettes 
 ## (Not recommended as vignettes contain the tutorials on using tidyqpcr)
-devtools::install_github("ewallace/tidyqpcr")
+devtools::install_github("ropensci/tidyqpcr")
 ```
 
 **Note**
@@ -95,6 +95,7 @@ Then load tidyqpcr as a standard package:
 ```
 library(tidyqpcr)
 ```
+
 **Note**
 tidyqpcr automatically imports and loads several external packages for basic functionality, including; tidy, dplyr and ggplot2.
 This allows tidyqpcr to be used immediately but may cause NAMESPACE clashes if the user already has many other package libraries loaded.
@@ -102,20 +103,20 @@ Restarting the R session and loading tidyqpcr separately may solve such issues.
 
 ## Using tidyqpcr
 
[…]
