### 2022-06-26 #4507 @arfon
@ewallace – I'm going to move to accept this submission in JOSS given the [earlier review](https://github.com/ropensci/software-review/issues/470) in rOpenSci.

Before I do this though could you:

- Take a look at fixing up the `10.1186/1471-2164-13-296/FIGURES/4` DOI? This looks to be invalid.
- Please make an archive of the software in Zenodo/figshare/other service and update this thread with the DOI of the archive? For the Zenodo/figshare archive, please make sure that:
