### 2022-06-26 #4507 @ewallace
Thank you @arfon!

- I fixed the DOI issue in https://github.com/ropensci/tidyqpcr/pull/195, thanks for finding that.
- tidyqpcr is now archived on Zenodo: https://doi.org/10.5281/zenodo.6757624
  - I checked the titles are the same
  - I checked the authors are the same

I hope that all works.
