### 2022-06-27 #4507 @arfon
@ewallace – thanks for making those changes. Your paper is now accepted and published in JOSS :zap::rocket::boom:
