### 2022-06-29 #4507 @ewallace
Hello @arfon! Thank you very much for your help.

Unfortunately we have an issue with the proof, that it still has a "DRAFT" watermark. How can we fix that?
